refactor(live-window): deduplicate utils, split WeatherLayer, standardize naming
Extract shared math utilities (clamp01, lerp, smoothstep, knuthHash) into
utils/math.ts. Consolidate color interpolation (rgbToHex, lerpColor, lerpHex)
into utils/color.ts, removing duplicates from MoonLayer, WeatherLayer, and
sky-gradient. Extract findPhasePosition to eliminate copy-pasted phase lookup
logic. Split 888-line WeatherLayer into weather/ submodules (types, configs,
colors, html). Rename blendGradient → lerpGradient for consistent naming.
